removes behaviour
